+++ example/example.dart
@@ -73,7 +73,7 @@ Future<void> main() async {
 
     // High-performance Volume Stream endpoint is available.
     final volumeStream = await twitter.tweetsService.connectVolumeStream();
-    await for (final response in volumeStream.handleError(print)) {
+    await for (final response in volumeStream.stream.handleError(print)) {
       print(response);
     }
 
@@ -87,7 +87,7 @@ Future<void> main() async {
     );
 
     final filteredStream = await twitter.tweetsService.connectFilteredStream();
-    await for (final response in filteredStream.handleError(print)) {
+    await for (final response in filteredStream.stream.handleError(print)) {
       print(response.data);
       print(response.matchingRules);
     }